requirements.txt
streamlit>=1.30.0
sounddevice>=0.4.6
numpy>=1.24.0
torch>=2.0.0
transformers>=4.36.0
openai>=1.6.0
python-dotenv>=1.0.0
scipy>=1.11.0  # Required for sounddevice
deepgram-sdk>=2.11.0 
onnxruntime>=1.20.1
kokoro-onnx
accelerate>=0.26.0
torchaudio